Moved functions from SharedResources to InterfaceFunctions; progress bar accounts for the extra vaccine analysis if present
